Miglioramenti grafici, diviso tutto il codice delle viste in file separati
