+++ server/ArtificialQI/API/views_def/__init__.py
@@ -1,0 +1,5 @@
+from .answer_views import *
+from .LLM_views import *
+from .prompt_views import *
+from .sessions_views import *
+from .test_views import *
